+++ src/react/components/fileuploader/private/FileList.js
@@ -120,7 +120,7 @@ const FileList = ({
                   <XUIButton
                     onClick={event => onRetry(file, fileList, event)}
                     size="small"
-                    variant="borderless-primary"
+                    variant="borderless-main"
                   >
                     {retryButtonText}
                   </XUIButton>
@@ -129,7 +129,7 @@ const FileList = ({
                   <XUIButton
                     onClick={event => handleCancel(file, event)}
                     size="small"
-                    variant="borderless-primary"
+                    variant="borderless-main"
                   >
                     {cancelButtonText}
                   </XUIButton>